Move filename parse method to utility module
